мелкий рефакторинг
