+++ linear-solver.html
@@ -7,7 +7,7 @@
     <script src="https://cdn.jsdelivr.net/combine/gh/paperjs/paper.js@v0.11.3/dist/paper-core.min.js,gh/jquery/jquery@2.2.4/dist/jquery.min.js,gh/kimmobrunfeldt/progressbar.js@1.0.1/dist/progressbar.min.js,gh/deepstreamIO/golden-layout@v1.5.8/dist/goldenlayout.min.js"></script>
     <!--script src="lib/genetic.js"></script>
     <script src="src/cutting.js"></script>
-    <script src="src/visualization.js"></script>
+    <script src="src/index.js"></script>
     <script src="src/linear-solver.js"></script-->
 
     <script src="dist/linear-solver.min.js"></script>
