+++ src/visualization/Visualization2D.js
@@ -1,0 +1,16 @@
+/**
+ *
+ *
+ * @module Visualization2D
+ *
+ * Created by Evgeniy Malyarov on 05.04.2020.
+ */
+
+class Visualization2D {
+
+  draw (userData, decision) {
+
+  }
+}
+
+module.exports = Visualization2D;
