+++ src/visualization/VisualizationPyramid.js
@@ -1,0 +1,16 @@
+/**
+ *
+ *
+ * @module VisualizationPyramid
+ *
+ * Created by Evgeniy Malyarov on 05.04.2020.
+ */
+
+class VisualizationPyramid {
+
+  draw (userData, decision) {
+
+  }
+}
+
+module.exports = VisualizationPyramid;
